AuthForm
created cone component for two pages SignIn and SignUp
