+++ src/pages/SigninPage/SigninPage.jsx
@@ -1,134 +1,43 @@
-import React, { useState } from 'react';
-import { Link } from 'react-router-dom';
+import { React, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { Formik, Form, Field, ErrorMessage } from 'formik';
-import * as Yup from 'yup';
-import { ReactComponent as OpenEyeIcon } from 'images/icons/eye-slash.svg';
-import { ReactComponent as ClosedEyeIcon } from 'images/icons/eye.svg';
+import { toast } from 'react-toastify';
+import Loader from 'components/Loader/Loader';
+import Backdrop from 'components/Backdrop/Backdrop';
 import { loginThunk } from '../../redux/auth/operations';
+import AuthForm from '../../components/AuthForm/AuthForm';
 import css from './SigninPage.module.css';
 
-const SigninPage = () => {
+const SignInPage = () => {
   const dispatch = useDispatch();
+  const [loader, setLoader] = useState(false);
 
-  const [showPassword, setShowPassword] = useState(false);
-
-  const initialValues = {
-    email: '',
-    password: '',
-  };
-
-  const validationSchema = Yup.object({
-    email: Yup.string().email().required('Email is required.'),
-    password: Yup.string()
-      .required('Password is required.')
-      .min(8, 'Password must be at least 8 characters.')
-      .max(64, 'Password must be less than 64 characters.'),
-  });
-
-  const onSubmit = (values, { resetForm }) => {
-    dispatch(loginThunk(values));
-    resetForm();
+  const signInHandler = (values, { resetForm }) => {
+    setLoader(true);
+    try {
+      dispatch(loginThunk(values));
+      toast.success('Sign in successful!');
+      resetForm();
+    } catch (error) {
+      setLoader(false);
+      toast.error('Sign in. Please try again.');
+    } finally {
+      setLoader(false);
+    }
   };
 
   return (
     <div className={css.container}>
       <div className={css.mainstr}>
         <div className={css.hidden}></div>
-        <Formik
-          initialValues={initialValues}
-          validationSchema={validationSchema}
-          onSubmit={onSubmit}
-        >
-          {({ errors, touched }) => (
-            <Form className={css.form}>
-              <h1 className={css.formTitle}>Sign In</h1>
-              <div className={css.formControl}>
-                <div className={css.stack}>
-                  <label className={css.formLabel} htmlFor="unique-id1">
-                    Enter your email
-                  </label>
-                  <Field
-                    id="unique-id1"
-                    className={`${css.input} ${
-                      errors.email && touched.email ? css.errorBorder : ''
-                    } ${errors.email && touched.email ? css.errorInput : ''}`}
-                    name="email"
-                    type="email"
-                    placeholder="Email"
-                  />
-                  <ErrorMessage
-                    name="email"
-                    component="div"
-                    className={css.errormessage}
-                  />
-                </div>
-
-                <div className={css.stack}>
-                  <label className={css.formLabel} htmlFor="unique-id2">
-                    Enter your password
-                  </label>
-                  <div
-                    className={`${css.inputBox} ${
-                      errors.password && touched.password ? css.errorBorder : ''
-                    }`}
-                  >
-                    <Field
-                      id="unique-id2"
-                      className={`${css.inputpassword}  ${
-                        errors.password && touched.password
-                          ? css.errorInput
-                          : ''
-                      }`}
-                      name="password"
-                      type={showPassword ? 'text' : 'password'}
-                      placeholder="Password"
-                    />
-                    <div
-                      className={css.iconeye}
-                      onClick={() => setShowPassword(!showPassword)}
-                      style={{
-                        cursor: 'pointer',
-                        marginTop: 'auto',
-                        marginBottom: 'auto',
-                      }}
-                    >
-                      {showPassword ? <ClosedEyeIcon /> : <OpenEyeIcon />}
-                    </div>
-                  </div>
-                  <ErrorMessage
-                    name="password"
-                    component="div"
-                    className={css.errormessage}
-                  />
-                </div>
-
-                <button className={css.button} type="submit">
-                  Sign In
-                </button>
-
-                <a
-                  href="https://aqua-backend-ieu7.onrender.com/auth/google"
-                  className={css.googleLogin}
-                >
-                  <p className={css.googleText}>Enter with Google</p>
-                </a>
-                <div className={css.wraplink}>
-                  <Link to="/signup" className={css.signup}>
-                    <p className={css.signupText}>Sign up</p>
-                  </Link>
-
-                  <Link to="/forgot-password" className={css.forgotPassword}>
-                    <p className={css.signupText}>Forgot password?</p>
-                  </Link>
-                </div>
-              </div>
-            </Form>
-          )}
-        </Formik>
+        <AuthForm formTitle="Sign In" onSubmit={signInHandler} />
+        {loader && (
+          <Backdrop>
+            <Loader />
+          </Backdrop>
+        )}
       </div>
     </div>
   );
 };
 
-export default SigninPage;
+export default SignInPage;
